requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
flake8==6.0.0
Flask==2.2.3
Flask-WTF==1.1.1
idna==3.4
imageio==2.24.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.1
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2022.7
requests==2.28.2
sqlparse==0.4.2
tzdata==2022.7
upgrade-pip==0.1.4
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
